FROMLIST: media: entity: Use pad as a starting point for graph walk
With the upcoming use of the recently added has_route() media entity op, all
the pads in an entity will no longer be considered interconnected. This has
an effect where the media graph is traversed: the starting pad does make a
difference.

Prepare for this change by using pad instead of the entity as an argument
for the graph walk operations. The actual graph traversal algorithm change
is in further patches.

Change-Id: I79b2d1d7734cbefe60fcf7055b7f4fb31e1d3038
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Reviewed-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Jacopo Mondi <jacopo+renesas@jmondi.org>
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
(cherry picked from https://lore.kernel.org/linux-media/20211005085750.138151-8-tomi.valkeinen@ideasonboard.com/)
Signed-off-by: Hsuan-Yu Lin <shane.lin@canonical.com>