ARM64: dts: mt8183: update gpu node to work with mali r20 driver
We have updated the mali driver.
This updates the gpu node to be compatible with the mali r20 driver.

Note:
The driver now support two regulators: mali and shadercores.
The mt8183 has two regulator the mali one and mali_sram.
Currently, we use set shadercores regulator for mali_sram
and assume that is right.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>
Signed-off-by: Hsuan-Yu Lin <shane.lin@canonical.com>