ANDROID: dma-heap: Provide accessors so that in-kernel drivers can allocate dmabufs from specific heaps
This allows drivers who don't want to create their own
DMA-BUF exporter to be able to allocate DMA-BUFs directly
from existing DMA-BUF Heaps.

There is some concern that the premise of DMA-BUF heaps is
that userland knows better about what type of heap memory
is needed for a pipeline, so it would likely be best for
drivers to import and fill DMA-BUFs allocated by userland
instead of allocating one themselves, but this is still
up for debate.

Unfortunately we don't have any public users of this, so
we cannot push it upstream. However, vendors have asked
for this, so I'm submitting it to Andorid Common in the
hopes that we can eventually find what vendors are doing.

[CPNOTE: 06/07/21] Lee: Asked vendor for an update via the bug

Signed-off-by: John Stultz <john.stultz@linaro.org>
Bug: 154341375

Signed-off-by: Yong Wu <yong.wu@mediatek.com>
Change-Id: I7b5ef690cefd2bfbbc2e01220ada49e44ef4fa04
Signed-off-by: Hsuan-Yu Lin <shane.lin@canonical.com>