FROMLIST: media: subdev: use streams in v4l2_subdev_link_validate()
Update v4l2_subdev_link_validate() to use routing and streams for
validation.

Instead of just looking at the format on the pad on both ends of the
link, the routing tables are used to collect all the streams going from
the source to the sink over the link, and the streams' formats on both
ends of the link are verified.

Change-Id: Ia6cbf1ee45119d64019858186511b0e06c3262b0
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
(cherry picked from https://lore.kernel.org/linux-media/20211005085750.138151-33-tomi.valkeinen@ideasonboard.com/)
Signed-off-by: Hsuan-Yu Lin <shane.lin@canonical.com>