iommu/mediatek: Change the domid to iova_region_id
Prepare for adding bankid, also no functional change.

In the previous SoC, each a iova_region is a domain; In the multi-banks
case, each a bank is a domain, then the original function name
"mtk_iommu_get_domain_id" is not proper. Use "iova_region_id" instead of
"domain_id".

Signed-off-by: Yong Wu <yong.wu@mediatek.com>
Link: https://lore.kernel.org/r/20210923115840.17813-30-yong.wu@mediatek.com
Signed-off-by: Hsuan-Yu Lin <shane.lin@canonical.com>