BL_INTERNAL: HACK: drm/mediatek: dpi: no power on/off in dpi_start/stop
Early power on/off in dpi_start/stop breaks MT8195 HDMI.
For the moment, we remove the power on/off calls first.

The dpi_power_on() is already called by bridge_enable().

However, the bridge_disable() does not do a full power_off(),
so this patch is still considered as a temoprary hack.

Signed-off-by: Guillaume Ranquet <granquet@baylibre.com>
Signed-off-by: Pablo Sun <pablo.sun@mediatek.com>
Signed-off-by: Hsuan-Yu Lin <shane.lin@canonical.com>