GENIO: drm/mediatek: Allow 2nd crtc to be optional
This patch fixes "DSI panel is not working when DPI0 is disabled".

In Genio 350 EVK, it is possible to disable DPI0 device in order to
enable ethernet pin mux. But disable DPI0 causes the crtc defined
by "mt8365_mtk_ddp_ext" fails to create, and the current logic aborts
and unbinds the entire DRM device, and makes DSI0 unavailable.

This patch relaxes the check of crtc creation, and allows "ext" and
"thrid" path to not creating crtc. The client could still use
the crtcs successfully created.

CR-Id: AUTO00204032
Change-Id: I61d15e316dfbe60c0c4ba9e21f592555cec45c33
Signed-off-by: Pablo Sun <pablo.sun@mediatek.com>
Signed-off-by: Hsuan-Yu Lin <shane.lin@canonical.com>