GENIO: arm64: dts: mediatek: genio-1200-evk: Update DMIC mode for PVT
On EVT, two DMICs were connected to DMIC1 and DMIC4, but on PVT
these two DMICs are connected to a single data pin of DMIC1. As
a result, the configuration for DMIC4 is no longer needed and has
been removed.

In addition, only one DMIC interface is remaining in use on PVT,
so the dmic-clk-mono mode and dmic-clk-index have been removed
from the configuration to simplify the setup.

This commit updates the DTS to reflect the hardware changes on PVT.

Change-Id: I813201ee736c9d0bc110fd110312e09fd25dc284
Signed-off-by: Zoran Zhan <zoran.zhan@mediatek.com>
Signed-off-by: Hsuan-Yu Lin <shane.lin@canonical.com>