GENIO: usb: typec: tcpm: tcpci_mt6360: fix tcpci_register_port failed log
tcpci_register_port may print an error log due to the late probing of
the Type-C mux driver. However, this log does not necessarily indicate
that the TCPC has really failed. After the Type-C mux is ready,
tcpci_mt6360 will probe again and succeed. Therefore,
the condition of printing this log will be modified in this situation.

Change-Id: Ic96a9319d141b1f52e1f14f7ee925b4199a20f92
Signed-off-by: Yow-Shin Liou <yow-shin.liou@mediatek.com>
Signed-off-by: Jian Hui Lee <jianhui.lee@canonical.com>
Signed-off-by: Hsuan-Yu Lin <shane.lin@canonical.com>