GENIO: mtk-mdp-driver: Prevent unreliable crash and driver clean up
[Problem]:
Once the m2m device is registered, there is a race between where
mdp->vpu_dev is populated and where it is accessed in
mtk_mdp_m2m_open().

[Solution]:
Move the v4l2_device_register and mtk_mdp_register_m2m_device to the end of the probe function to ensure that the device registration gets done only after all driver elements are populated.

This MR also cleans up the error logging in the mdp driver.

Change-Id: I24800ffd9d1a2e8912c35504a033413d23ca11f8
Signed-off-by: Macross Chen <macross.chen@mediatek.com>
Signed-off-by: Suhrid Subramaniam <suhrid.subramaniam@mediatek.com>