wifi: iwlwifi: mvm: fix recovery flow in CSA
commit 828c79d9feb0 ("wifi: iwlwifi: mvm: fix recovery flow in CSA") upstream.

If the firmware crashes in the de-activation / re-activation
of the link during CSA, we will not have a valid phy_ctxt
pointer in mvmvif. This is a legit case, but when mac80211
removes the station to cleanup our state during the
re-configuration, we need to make sure we clear ap_sta
otherwise we won't re-add the station after the firmware has
been restarted. Later on, we'd activate the link, try to send
a TLC command crash again on ASSERT 3508.

Fix this by properly cleaning up our state.

deepin-Intel-SIG: commit 828c79d9feb0 ("wifi: iwlwifi: mvm: fix recovery flow in CSA").

Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Signed-off-by: Gregory Greenman <gregory.greenman@intel.com>
Link: https://lore.kernel.org/r/20230913145231.2651e6f6a55a.I4cd50e88ee5c23c1c8dd5b157a800e4b4c96f236@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
[ Furong Zhou: amend commit log ]
Signed-off-by: Furong Zhou <furong.zhou@intel.com>