wifi: iwlwifi: mvm: simplify the reorder buffer
commit ff8e3a40d78b ("wifi: iwlwifi: mvm: simplify the reorder buffer") upstream.

The firmware / hardware of devices supporting RSS is able to report
duplicates and packets that time out inside the reoder buffer. We can
now remove all the complex logic that was implemented to keep all the Rx
queues more the less synchronized: we used to send a message to all the
queues through the firmware to teach the different queues about what is
the current SSN every 2048 packets.

Now that we rely on the firmware / hardware to detect duplicates, we can
completely remove the code that did that in the driver and it has been
reported that this code was spuriously dropping legit packets.

deepin-Intel-SIG: commit ff8e3a40d78b ("wifi: iwlwifi: mvm: simplify the reorder buffer").

Suggested-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Signed-off-by: Gregory Greenman <gregory.greenman@intel.com>
Link: https://lore.kernel.org/r/20231017115047.54cf4d3d5956.Ic06a08c9fb1e1ec315a4b49d632b78b8474dab79@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
[ Furong Zhou: amend commit log ]
Signed-off-by: Furong Zhou <furong.zhou@intel.com>