UBUNTU: SAUCE: (no-up) GENIO: drm/mediatek: bridge: it6122: add usage of power regulator
The it6122 bridge power shares the pwr_en GPIO pin with
panel-koe-tx18d204vm0baa. To allow both the it6122 and the panel drivers
to enable the power separately, utilize a GPIO regulator.

Add the power regulator usage flow as follows:

1. in it6122_power_on_off
Enable the regulator for the 'on' case and disable it for the 'off' case.
This function is called when suspending or resuming to power off or on
the it6122.

2. in it6122_probe
Enable the regulator to read the chip ID and then disable it after the
read operation. If the regulator is not disabled in it6122_probe, it will
be enabled again in the subsequent it6122_power_on_off call with the 'on'
case, and then it will remain enabled during suspend because the
it6122_power_on_off call with the 'off' case will only disable it once.

3. modify it6122_bridge_enabled to it6122_bridge_probed
Since it6122_probe disables the power regulator, this function actually
returns if the it6122 is probed, rather than enabled.

Change-Id: I5e423fd8ca2f5617aa80347c146531c066064d50
Signed-off-by: Tommy Chen <tommyyl.chen@mediatek.com>
Signed-off-by: Jian Hui Lee <jianhui.lee@canonical.com>