UBUNTU: SAUCE: (no-up) mt76: duplicate the driver for oot updates
MediaTek would like to freeze the code base of mt76 after patching the
driver from their verified repo, so put this driver into the ubuntu
folder for the maintenance.
Since the in-kernel mt76 driver is not going to be maintained by
MediaTek, so disable it and point to the out-of-tree mt76 driver under
ubuntu directory.

The kernel version for the mt76 driver base is Ubuntu-5.15.0-46.49.

Signed-off-by: Jian Hui Lee <jianhui.lee@canonical.com>