UBUNTU: SAUCE: (no-up) GENIO: media: i2c: ap1302: Poll BOOTDATA_STAGE after loading bootdata
This commit introduces the addition of bootdata stage polling
after loading the bootdata content is done. Such an installation
has become indispensable to ensure that the bootdata stage in
the ap1302 hardware completes. Without this crucial sync process,
there's a potential risk of a race condition arising, thereby
inducing a discordance between the software and hardware status.

Change-Id: I0cf5c877e9b4efb9aa263643fc9e828adea07d39
Signed-off-by: Andy.Hsieh <andy.hsieh@mediatek.com>
Signed-off-by: Jian Hui Lee <jianhui.lee@canonical.com>