GPUCORE-41385 Fix small memory aliases of large memory allocations
Previously, when creating a GPU mapping, the driver would only consider
the alignment of the GPU VA and the tag attached to the corresponding
physical page to decide whether to create a single MMU ATE at level 2
for a 2 MB region or a set of MMU ATEs at level 3 of 4 kB each.

The problem was that a small memory alias of a large memory allocation
could satisfy both conditions if its GPU VA happens to be aligned to
a large memory page. This would lead to the creation of a 2 MB mapping
even if the user only requested a smaller region, e.g. 4 kB. Moreover,
this would lead to problem during the teardown, leaving the MMU page
table in an inconsistent state and potentially keeping the GPU mapping
in existence, allowing use-after-free of a physical page after the
original allocation is released.

The driver now adds a third condition to the creation of a MMU ATE
at level 2: the total number of MMU page table entries to map must be
512, i.e. the user has requested a whole 2 MB region intentionally.

A defect test has been introduced to reproduce the problem. The test
creates a memory alias with 513 members of 4 kB each, which reproduces
the conditions which are necessary to trigger the bug. After that,
it frees the aliases and attempts a GPU write after free.

This fix addresses CVE-2024-0671 ("Use-after-free: dangling ATE via small alias of 2MiB page").

CR-Id: AUTO00301959
Change-Id: Idd9601b179ed28e7bd91be7259a2ad70ef4567dd
Signed-off-by: Guoqing Jiang <guoqing.jiang@canonical.com>
Signed-off-by: Jian Hui Lee <jianhui.lee@canonical.com>