GPUCORE-41547 Improve handling of page metadata after migration
Handling of page metadata after migration in kbase_mmu_migrate_page()
has been made slightly safer. Only if the page is still in the
ALLOCATED_MAPPED state is it safe to access and update the
metadata. There's no problem if the page became NOT_MOVABLE.
It is an error, instead, if the page has changed to any other
state.

The kbase_set_phy_alloc_page_status() helper must be called
while keeping the region lock to prevent race conditions where
the pages of a GPU VA region could be set to NOT_MOVABLE
while a page migration is in progress.

Change-Id: Ie248adee117fbb93c16201926c25baafae2b0ee3
(Cherry-picked for CVE-2024-4607)
CR-Id: AUTO00309986
Signed-off-by: Pablo Sun <pablo.sun@mediatek.com>
Signed-off-by: Guoqing Jiang <guoqing.jiang@canonical.com>
Signed-off-by: Jian Hui Lee <jianhui.lee@canonical.com>