UBUNTU: [Config] CONFIG_DRM_PANEL_STARTEK_KD070FHFID015=m
BugLink: http://bugs.launchpad.net/bugs/2069787 ([UC22][g350] systemd-backlight@backlight:14014000.dsi0.0.service: Failed to write system 'brightness' attribute: Invalid argument during warm-boot / cold-boot stress test (LP: #2069787))

on g350, it66121(CONFIG_DRM_ITE_IT66121) is the bridge ic, which the
driver is built as a module. with the built-in panel driver, the
backlight registration would happen much eariler, but not usable. it
needs to wait until the bridge ic is loaded and enables the panel, which
would cause a time gap, especially on ubuntu core system.

making the panel driver into module is a reasonable change because when
the panel is being probing and attaching to dsi, it would wake up the
parent and makes it to enable the panel.

Signed-off-by: Jian Hui Lee <jianhui.lee@canonical.com>