UBUNTU: SAUCE: (no-up) GENIO: panel: startek-kd070fhfid015: change enable gpio to be optional
BugLink: https://bugs.launchpad.net/bugs/2081699

On G510 EVK, the startek-kd070fhfid078 (abbreviated as kd-078) utilizes
GPIO 111 as a gpio-regulator. However, on legacy G700 EVK, the
startek-kd070fhfid015 (abbreviated as kd-015) uses GPIO pin 111 as an
enable-gpio.

On the latest G700 EVK p1v4, kd-078 is used instead. In order to make
the G700 EVK dts compatible with both kd-015 and kd-078, GPIO 111 is
modified to be a gpio-regulator. It now acts as the iovcc regulator for
kd-015, which causes the kd-015 probe to fail as it can not find
enable-gpio during the probe process.

As a result, the enable-gpio in the kd-015 driver is made optional.
The G350 EVK still use this enable-gpio pin.

Change-Id: I197fd74309f5cb1c54e951a40aed7ec4657ad537
Signed-off-by: Tommy Chen <tommyyl.chen@mediatek.com>
Signed-off-by: Jian Hui Lee <jianhui.lee@canonical.com>