GENIO: ASoC: mediatek: mt6359-accdet: Add compatible property
Add compatible property for mt6359-accdet platform driver.
Add accdet_of_match, this of_match_table provides a mapping between
the compatible property in the device tree and the platform driver.
Without this table, the kernel would not be able to associate the
mt6359-accdet driver with the corresponding hardware.

Change-Id: I3646f17215c72dd08e0d42ef7acaa60396021cc6
Signed-off-by: Andrew Perepech <andrew.perepech@mediatek.com>
Signed-off-by: Guoqing Jiang <guoqing.jiang@canonical.com>
Signed-off-by: Jian Hui Lee <jianhui.lee@canonical.com>