UBUNTU: [Config] CONFIG_DRM_MEDIATEK_HDMI=y and CONFIG_PHY_MTK_HDMI=y
BugLink: http://bugs.launchpad.net/bugs/2078763

on g700, there may be a gap between the panel being probed and fully enabled
(prepared). this gap would cause the systemd-backlight service to fail
because the backlight sysfs entry is created during the probing stage.

to fix this gap in kernel, we need to make mediatek hdmi module
built-in. this way, so when the panel is being probed, it will attempt to wake up the
master, which is mediatek drm. since every binded components are
already added, the gap will be neglected.

Signed-off-by: Jian Hui Lee <jianhui.lee@canonical.com>