GENIO: media: i2c: ap1302: Set regmap name
Fix the following error log by setting the name in regmap_config. The
ap1302 driver registers the regmap twice because the ap1302 register
has two access modes, one is 16-bit and the other is 32-bit. If the
name is empty, the regmap will create the node named '2-003d' twice
leading to the kernel error.

Error log:

  [    4.009081] debugfs: Directory '2-003d' with parent 'regmap' already present!

After setting the name, the regmap debugfs nodes will be:

  /sys/kernel/debug/regmap/2-003d-32bit
  /sys/kernel/debug/regmap/2-003d-16bit

Change-Id: Ia50f4413180b6ef2a3889387cb4c69e76c031d97
Signed-off-by: Andy.Hsieh <andy.hsieh@mediatek.com>