GENIO: drm/mediatek: dp: Remove DP enabled check
This commit removes the check for DP enabled status in the
mtk_dp_audio_hw_params function. The original check caused issues on
Ubuntu systems when the physical DP device was not connected to the
board before powering on, which led to audio system failure.

By removing this check, we allow the DP audio configuration to proceed
even if the physical DP device is not connected to the board when the
system is powered on.

Change-Id: I6d4577c3c8159f0503cac8aa9ed5ce92d91cdab3
Signed-off-by: Zoran Zhan <zoran.zhan@mediatek.com>