UBUNTU: SAUCE: (no-up) power: Ignore duplicate OPP entry to fix probe fail
BugLink: https://bugs.launchpad.net/bugs/2090978

On Genio 510 (mt8370), the OPP entries of DLA hardware
has duplicated entries on rate 700000000.

This is caused by the design of the internal implementation
of APUSYS driver. The duplicated entries are used to
match with other hardware such as VPU, and the number of
OPP entries must be the same. Hence the duplicated entries.

On Genio 510, the devfreq device fails to probe with the
following error:

[ 13.668520] [mt8188_apu_devfreq_cooling_register][238] dev_name(&pdev->dev) = soc:mdla_devfreq, dvfs_dev_id = 1
[ 13.668650] mt8188-mdla-devfreq soc:mdla_devfreq: _opp_is_duplicate: duplicate OPPs detected. Existing: freq: 700000000, volt: 800000, enabled: 1. New: freq: 700000000, volt: 750000, enabled: 1
[ 13.668658] [mt8188_apu_devfreq_cooling_register][296] dvfs_dev_id = 1, register cooling dev NG!
[ 13.668662] mt8188-mdla-devfreq: probe of soc:mdla_devfreq failed with error -17

Since this is expected and it does not affect the logic of
the actual frequency operations in APUSYS, we ignore the
EEXIST error and continue the probing process.

Also adjust the order of debug log, so the warning log
appears after the debug log.

Bug: https://bugs.launchpad.net/baoshan/+bug/2069663
Change-Id: I12385308aab43bc30587d5691802ee9d54d7c042
Signed-off-by: Pablo Sun <pablo.sun@mediatek.com>
(cherry picked from commit 6b2de9b241cfa23531c8ceaf88ac80fbe411ed8e
https://gitlab.com/mediatek/aiot/bsp/mtk-apusys-driver)