GENIO: panel: startek-kd070fhfid015: fix setting brightness failure during early stage booting
The original idea was to make the panel driver into a module, and when the
panel is being probed and attaching to DSI, this would wake up the
parent, which is DSI, and enable the panel.

To achieve that, the DSI module must be probed earlier than the panel
driver. However, there is still a chance that the DSI module could be
probed after the panel driver, which means that the upper layer could
update the panel brightness before the panel is prepared.

Returning EINVAL when this situation happens would cause upper layer to
show error. So return success instead and later restore the brightness
during stk_panel_init(), as the value from the upper layer is already
saved in the backlight structure.

Change-Id: If67ae5b92696ba272648e4c6773450e9199766b1
Signed-off-by: Jian Hui Lee <jianhui.lee@canonical.com>
Signed-off-by: Tommy Chen <tommyyl.chen@mediatek.com>
Signed-off-by: Jian Hui Lee <jianhui.lee@canonical.com>