GENIO: drm/mediatek: panel: add drm_mode_set_name to AUO G156XTN010 LVDS panel
When running modetest on the AUO G156XTN010 LVDS panel with the command

modetest -M mediatek -s 32@43:1366x768

it can not be executed because the mode 1366x768 is not set as a valid
name. As a result, add drm_mode_set_name in panel_lvds_get_modes of
the AUO G156XTN010 LVDS panel driver.

Change-Id: If94c9592b30b0da80d96a04c4255e6f95fb4d9dc
CR-Id: AUTO00351838
Signed-off-by: Tommy Chen <tommyyl.chen@mediatek.com>
Signed-off-by: Guoqing Jiang <guoqing.jiang@canonical.com>
Signed-off-by: Jian Hui Lee <jianhui.lee@canonical.com>