dtbo: apusys.dts: Set always-on for apusys power supplies
Set APU power supplies to always-on to mitigate an issue
that causes APU fail to power on in certain Genio 700 (MT8370)
boards. Also update Genio 510 dtbo for consistency.

The APU may fail to power on when it is power cycled repeatedly,
and the following middleware 'Inference fail' messages appear:

[apusys][error]submit: run cmd fail(-108/Cannot send after transport endpoint shutdown)
ERROR: APUSysEngine::RunWithTypeCmd() failed
ERROR: A work (CompiledResult) index = 0 failed in parallel execution.
ERROR: RunCompiledResultParallel failed
ERROR: Runtime inference fails. Runtime inference error.
ERROR: There is a runtime error during inference
ERROR: Inference fail

This issue could not be reproduced on Genio 700 EVK; But it is known
to occur on certain hardware boards. Some boards have much higher
reproducibility than other boards.

In reference Android design, both power supplies are flagged
as always-on. Follow the Android design and this issue cannot be
reproduced on any of the boards.

Note that on Android reference design, the APU power supplies,
although marked as always-on, are actually controlled by the
APU firmware. This is controlled by the configration option
ENABLE_SW_BUCK_CTL in both trusted-firmware-a and apusys kernel
driver.

On reference Android, the option ENABLE_SW_BUCK_CTL is set to 0.
In IoT Yocto v24.0 and before, it is set to 1.

As a result, the idle power of Genio 700 EVK increases
after applying this patch, if ENABLE_SW_BUCK_CTL remains enabled.

Note that the idle power of Genio 510 EVK doesn't increase
because its board DTS already set power supplies always-on[1].

The patches to set ENABLE_SW_BUCK_CTL to 0 in TF-A and apusys driver
is submitted separately.

[1]: https://gitlab.com/mediatek/aiot/bsp/linux/-/commit/7021ac6d97b79d5725f56f0fc57404d63f698e87
Change-Id: Iabd600150822f5008ef465d7df1ffb9d1be20349
CR-Id: ALPS09375096
Signed-off-by: Pablo Sun <pablo.sun@mediatek.com>
Signed-off-by: Jian Hui Lee <jianhui.lee@canonical.com>